Move login to separate path
