+++ routes/api-router.js
@@ -3,6 +3,7 @@ const { methodNotAllowed } = require('../errors');
 const usersRouter = require('./users-router');
 const sitesRouter = require('./sites-router');
 const pinsRouter = require('./pins-router');
+const loginRouter = require('./login-router');
 
 apiRouter
   .route('/')
@@ -14,5 +15,7 @@ apiRouter.use('/users', usersRouter);
 apiRouter.use('/sites', sitesRouter);
 
 apiRouter.use('/pins', pinsRouter);
+
+apiRouter.use('/login', loginRouter);
 
 module.exports = apiRouter;
