+++ routes/login-router.js
@@ -1,0 +1,10 @@
+const loginRouter = require('express').Router();
+const { methodNotAllowed } = require('../errors');
+const { authenticateUser } = require('../controllers/login-controller');
+
+loginRouter
+  .route('/')
+  .post(authenticateUser)
+  .all(methodNotAllowed);
+
+module.exports = loginRouter;
