+++ routes/users-router.js
@@ -1,7 +1,6 @@
 const usersRouter = require('express').Router();
 const { methodNotAllowed } = require('../errors');
 const {
-  authenticateUser,
   getAllUsers,
   getUser,
   postUser,
@@ -13,11 +12,6 @@ usersRouter
   .route('/')
   .get(getAllUsers)
   .post(postUser)
-  .all(methodNotAllowed);
-
-usersRouter
-  .route('/login')
-  .post(authenticateUser)
   .all(methodNotAllowed);
 
 usersRouter
